skillering-app/python/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.9
python-dotenv
pandas
google-generativeai
google-auth-httplib2
google-api-core
google-api-python-client

